refactor: extract features list details from pricing section into a reusable component.
